LPS-152783 Simplify
