+++ modules/apps/site-initializer/site-initializer-testray/extra/remote-app/src/i18n/en_US.ts
@@ -51,6 +51,7 @@ export default {
 	'build': 'Build',
 	'build-history': 'Build History',
 	'build-name': 'Build Name',
+	'cancel': 'Cancel',
 	'case': 'Case',
 	'case-details': 'Case Details',
 	'case-name': 'Case Name',
